add lizardfs benchmark

+++ crcutil.cc
@@ -1,12 +1,12 @@
 #include <stdint.h>
 #include "crcutil/generic_crc.h"
 
-extern "C" unsigned long crc32(uint32_t prev, void  * data, size_t length);
+extern "C" uint32_t crc32(uint32_t prev, void  * data, size_t length);
 
-static constexpr int32_t CRC_POLY = 0xEDB88320U;
+static constexpr uint32_t CRC_POLY = 0xEDB88320U;
 
 static crcutil::GenericCrc<uint32_t, uint32_t, size_t, 4> gCrc(CRC_POLY, 32, true);
 
-unsigned long crc32(uint32_t prev, void * data, size_t length){
+uint32_t crc32(uint32_t prev, void * data, size_t length){
   return gCrc.CrcDefault(data, length, prev);
 }
